+++ docs/changelog.rst
@@ -5,6 +5,7 @@ Changelog
 
 Changes in 0.8.2
 ================
+- Fixed cascading saves which weren't turned off as planned (#291)
 - Fixed Datastructures so instances are a Document or EmbeddedDocument (#363)
 - Improved cascading saves write performance (#361)
 - Fixed amibiguity and differing behaviour regarding field defaults (#349)